Create README.md
# Db_seeder
This short code will connect to mysql server and create a database. Then it will create a table inside it and populate it with 1mil entries. 
we will then select the first row from the database and compae the strings to find the largest string i.e. the string sith mire ccharacters. If we get more than one string then we will mention the number of those strings.